기본 예외/로그 처리
기본 예외/로그 처리

+++ src/main/java/com/kit/pillgood/service/PrescriptionService.java
@@ -1,11 +1,14 @@
 package com.kit.pillgood.service;
 
+import com.kit.pillgood.exeptions.exeption.NonExistsPrescriptionIndexException;
 import com.kit.pillgood.exeptions.exeption.NonRegistrationGroupException;
 import com.kit.pillgood.persistence.dto.PrescriptionAndDiseaseNameDTO;
 import com.kit.pillgood.persistence.projection.PrescriptionAndDiseaseNameSummary;
 import com.kit.pillgood.repository.GroupMemberRepository;
 import com.kit.pillgood.repository.PrescriptionRepository;
 import com.kit.pillgood.util.EntityConverter;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -15,6 +18,7 @@ import java.util.List;
 
 @Service
 public class PrescriptionService {
+    private final Logger LOGGER = LoggerFactory.getLogger(PrescriptionService.class);
     private final PrescriptionRepository prescriptionRepository;
     private final GroupMemberRepository groupMemberRepository;
 
@@ -41,11 +45,18 @@ public class PrescriptionService {
         for(PrescriptionAndDiseaseNameSummary prescriptionAndDiseaseNameSummary : prescriptionAndDiseaseNameSummaries) {
             prescriptionAndDiseaseNameDTOs.add(EntityConverter.toPrescriptionAndDiseaseNameDTO(prescriptionAndDiseaseNameSummary));
         }
+
+        LOGGER.info("처방전 조회 성공 {}", prescriptionAndDiseaseNameDTOs);
         return prescriptionAndDiseaseNameDTOs;
     }
 
-    public void deletePrescription(Long prescriptionIndex) {
+    public void deletePrescription(Long prescriptionIndex) throws NonExistsPrescriptionIndexException {
+        if(!prescriptionRepository.existsByPrescriptionIndex(prescriptionIndex)){
+            LOGGER.info("[err] prescriptionIndex={}을 찾을 수 없음", prescriptionIndex);
+            throw new NonExistsPrescriptionIndexException();
+        }
         prescriptionRepository.deleteById(prescriptionIndex);
+        LOGGER.info("[err] prescriptionIndex={} 삭제 성공", prescriptionIndex);
     }
 
 }
